Move event arguments into separate files
